requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
alembic
qdrant-client
psycopg2-binary
python-json-logger
pymupdf
python-multipart
